+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2020-02-27-16:47 - gearysw
+
+## Features
+- First pass of Google Sheets API token refresh logic pushed
+
 # 2020-02-27 - gearysw
 
 ## Features